+++ crates/serialization/src/lib.rs
@@ -43,8 +43,8 @@
 //! or as part of a Rust only build.
 //!
 //! - `arrow`: Enables Apache Arrow schema definitions and RecordBatch encoding/decoding.
+//! - `arrow-display`: Enables display-friendly Arrow encoders for market data and requires `arrow`.
 //! - `capnp`: Enables [Cap'n Proto](https://capnproto.org) serialization support.
-//! - `display`: Enables display-friendly Arrow encoders for market data and requires `arrow`.
 //! - `extension-module`: Builds as a Python extension module.
 //! - `high-precision`: Enables
 //!   [high-precision mode](https://nautilustrader.io/docs/nightly/getting_started/installation/#precision-mode)